chore: globalstyle 및 preview decorator 적용
