+++ src/styles/global.tsx
@@ -1,0 +1,3 @@
+import { css } from "@emotion/react";
+
+export const globalStyles = css``;
